split admin panel
