+++ tgui/packages/tgui/interfaces/PlayerEffects/PlayerEffectsTabs/ControlFixes.tsx
@@ -1,0 +1,20 @@
+import { useBackend } from '../../../backend';
+import { Button, Section } from '../../../components';
+
+export const ControlFixes = (props) => {
+  const { act } = useBackend();
+
+  return (
+    <Section title="Bug Fixes">
+      <Button fluid onClick={() => act('rejuvenate')}>
+        Rejuvenate
+      </Button>
+      <Button fluid onClick={() => act('popup-box')}>
+        Send Message Box
+      </Button>
+      <Button fluid onClick={() => act('stop-orbits')}>
+        Clear All Orbiters
+      </Button>
+    </Section>
+  );
+};
