+++ tgui/packages/tgui/interfaces/PlayerEffects/types.ts
@@ -1,0 +1,4 @@
+export type Data = {
+  real_name: string;
+  player_ckey: string;
+};
